moved copingSkillCard and updated imports accordingly
